Refactor: moved implementation to module types. Created new module users.
