--- src/main/java/com/lasalle/sd2/g2/domain/PokemonRepository.java
@@ -1,6 +1,0 @@
-package com.lasalle.sd2.g2.domain;
-
-public interface PokemonRepository {
-
-    PokemonTypes getPokemonTypes(String pokemonName);
-}
