--- src/main/java/com/lasalle/sd2/g2/domain/PokemonTypesException.java
@@ -1,10 +1,0 @@
-package com.lasalle.sd2.g2.domain;
-
-public class PokemonTypesException extends Exception {
-
-    private static final long serialVersionUID = -9196422863982068798L;
-
-    public PokemonTypesException(String message) {
-        super(message);
-    }
-}
