--- src/main/java/com/lasalle/sd2/g2/infrastructure/dto/ErrorMessageDto.java
@@ -1,14 +1,0 @@
-package com.lasalle.sd2.g2.infrastructure.dto;
-
-public class ErrorMessageDto {
-
-    private final String message;
-
-    public ErrorMessageDto(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
