+++ src/main/java/com/lasalle/sd2/g2/infrastructure/server/JettyServer.java
@@ -1,7 +1,7 @@
 package com.lasalle.sd2.g2.infrastructure.server;
 
 import com.lasalle.sd2.g2.infrastructure.conf.AppProperties;
-import com.lasalle.sd2.g2.infrastructure.controller.PokemonTypesServlet;
+import com.lasalle.sd2.g2.types.infrastructure.controller.PokemonTypesServlet;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.eclipse.jetty.server.Connector;
