+++ types/src/main/java/com/lasalle/sd2/g2/types/domain/PokemonRepository.java
@@ -1,0 +1,6 @@
+package com.lasalle.sd2.g2.types.domain;
+
+public interface PokemonRepository {
+
+    PokemonTypes getPokemonTypes(String pokemonName);
+}
