+++ types/src/main/java/com/lasalle/sd2/g2/types/domain/PokemonTypesException.java
@@ -1,0 +1,10 @@
+package com.lasalle.sd2.g2.types.domain;
+
+public class PokemonTypesException extends Exception {
+
+    private static final long serialVersionUID = -9196422863982068798L;
+
+    public PokemonTypesException(String message) {
+        super(message);
+    }
+}
