+++ types/src/main/java/com/lasalle/sd2/g2/types/infrastructure/conf/TypesConfiguration.java
@@ -1,0 +1,14 @@
+package com.lasalle.sd2.g2.types.infrastructure.conf;
+
+public class TypesConfiguration {
+
+    private static String pokemonTypesBaserUrl;
+
+    public static String getBaseUrl() {
+        return pokemonTypesBaserUrl;
+    }
+
+    public static void setBaseUrl(String pokeApiBaserUrl) {
+        TypesConfiguration.pokemonTypesBaserUrl = pokeApiBaserUrl;
+    }
+}
